Persist some data between protocol server restarts
Sometimes the udp connection to the broker dies, and the protocol server
needs to reconnect. In those cases, the whole applications doesn't die, and
clients registered as listeners should not have to re-register. Thus,
information about listeners and which topics to subscribe to are stored
in a separate process, to keep state between protocol server restarts.

If the whole application restarts, the state will be lost.